take over SPI bus completely during disk transactions
just a test, but it seems to help a lot. needs a cleaner implementation

+++ lib/v06x/src/8253.h
@@ -10,10 +10,10 @@ class CounterUnit
 {
     friend class TestOfCounterUnit;
 
-    int latch_value;
-    int write_state;
-    int latch_mode;
-    int mode_int;
+    uint16_t latch_value;
+    uint8_t write_state;
+    uint8_t latch_mode;
+    uint8_t mode_int;
 
     uint8_t write_lsb;
     uint8_t write_msb;
@@ -31,7 +31,7 @@ public:
         };
     };
 
-    int out;
+    uint8_t out;
     uint16_t loadvalue;
     int value;
 
@@ -60,8 +60,8 @@ private:
     CounterUnit counters[3];
     uint8_t control_word;
     //int clock_carry;
-    int counted_carry;
-    int accu_carry;
+    int16_t counted_carry;
+    int16_t accu_carry;
 
 public:
     audio_sample_t * audio_buf; // pointer to external sound buf
